[gfr] `Config` is (partially) applied (`--nr-fun-disks`).
The names of the "fun" disks have been changed
from `Fun0`, `Fun1` etc. to `Fun_00h`, `Fun_01h` etc.

Even though the numbering in the names is now in hexadecimal,
the value for `--nr-fun-disks` still has to be supplied in decimal.
TODO: We might want to allow the User to use hexadecimal in the cmdline.